# unrestrict chunk size