--- a/docs/advanced_usage.html
+++ b/docs/advanced_usage.html
@@ -252,11 +252,9 @@ by using the in-place API functions:</p>
 &#8220;perform your own benchmarks and see what works best for your use case.&#8221; As stated before,
 the optimal chunk size is heavily dependent on the target function and length of the input
 vector.</p>
-<p>When selecting a chunk size, keep in mind that the maximum allowed size is <code class="docutils literal"><span class="pre">10</span></code> (to
-change this, you can alter the <code class="docutils literal"><span class="pre">MAX_CHUNK_SIZE</span></code> constant in ForwardDiff&#8217;s source and
-reload the package). Also, it is usually best to pick a chunk sizes which divides evenly
-into the input dimension. Otherwise, ForwardDiff has to construct and utilize an extra
-&#8220;remainder&#8221; chunk to complete the calculation.</p>
+<p>Note that it is usually best to pick a chunk size which divides evenly into the input
+dimension. Otherwise, ForwardDiff has to construct and utilize an extra &#8220;remainder&#8221; chunk to
+complete the calculation.</p>
 </div>
 <div class="section" id="hessian-of-a-vector-valued-function">
 <h2>Hessian of a vector-valued function<a class="headerlink" href="#hessian-of-a-vector-valued-function" title="Permalink to this headline">¶</a></h2>