--- a/test/utils.jl
+++ b/test/utils.jl
@@ -1,11 +1,11 @@
 import ForwardDiff
-using ForwardDiff.MAX_CHUNK_SIZE
+using ForwardDiff.CHUNK_THRESHOLD
 using Base.Test
 
-const XLEN = MAX_CHUNK_SIZE
-const YLEN = div(MAX_CHUNK_SIZE, 2) + 1
+const XLEN = CHUNK_THRESHOLD
+const YLEN = div(CHUNK_THRESHOLD, 2) + 1
 const X, Y = rand(XLEN), rand(YLEN)
-const CHUNK_SIZES = (1, div(MAX_CHUNK_SIZE, 3), div(MAX_CHUNK_SIZE, 2), MAX_CHUNK_SIZE)
+const CHUNK_SIZES = (1, div(CHUNK_THRESHOLD, 3), div(CHUNK_THRESHOLD, 2), CHUNK_THRESHOLD)
 const FINITEDIFF_ERROR = 1e-5
 
 # used to test against results calculated via finite difference