… docs/searchindex.js