A while ago, I refactored ForwardDiff to remove all `@generated` functions from the codebase. A direct effect of this was that a bunch of elementwise functions on `Partials` had to be restricted to a small subset of chunk sizes (from `1` to `10`). Obviously, this restriction can be annoying for users who want to play around with chunk sizes greater than `10`. This PR reintroduces `@generated` functions for the affected `Partials` code, thus enabling the use of any chunk size via the API.

AFAICT after a small bit of profiling, this doesn't affect performance at all. I'll do some more benchmarking later today to confirm that, though.

Note also that I intend to tag ForwardDiff v0.3.0 after this PR gets merged.