@KristofferC 2016-11-29
+1. I have found that 10 chunk size is a bit small.