@jrevels 2016-11-29
> How's the hit on compilation time?

There's a slight slowdown, but it doesn't seem too significant. Also, I think #137 might be at least somewhat fixed on the `master` branch of ForwardDiff now, so I suppose it's not as much of an issue?

Here's the code I'm using to check this:

```julia
using ForwardDiff, BenchmarkTools

function rosenbrock(x)
   a = one(eltype(x))
   b = 100 * a
   result = zero(eltype(x))
   for i in 1:length(x)-1
       result += (a - x[i])^2 + b*(x[i+1] - x[i]^2)^2
   end
   return result
end

out, x = zeros(100, 100), rand(100);

cfg = ForwardDiff.HessianConfig(x);

# measures compilation performance
@time ForwardDiff.hessian!(out, rosenbrock, x, cfg);

# measures runtime performance
@benchmark ForwardDiff.hessian!($out, $rosenbrock, $x, $cfg) evals=1
```

On master (showing the two lines we care about):

```julia
julia> @time ForwardDiff.hessian!(out, rosenbrock, x, cfg);
  2.081374 seconds (2.03 M allocations: 45.544 MB, 1.04% gc time)

julia> @benchmark ForwardDiff.hessian!($out, $rosenbrock, $x, $cfg) evals=1
BenchmarkTools.Trial:
  memory estimate:  87.52 kb
  allocs estimate:  11
  --------------
  minimum time:     5.263 ms (0.00% GC)
  median time:      5.271 ms (0.00% GC)
  mean time:        5.315 ms (0.13% GC)
  maximum time:     8.135 ms (33.59% GC)
  --------------
  samples:          941
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

On this branch:

```julia
julia> @time ForwardDiff.hessian!(out, rosenbrock, x, cfg);
  2.188665 seconds (2.12 M allocations: 48.981 MB, 0.84% gc time)

julia> @benchmark ForwardDiff.hessian!($out, $rosenbrock, $x, $cfg) evals=1
BenchmarkTools.Trial:
  memory estimate:  87.52 kb
  allocs estimate:  11
  --------------
  minimum time:     5.265 ms (0.00% GC)
  median time:      5.274 ms (0.00% GC)
  mean time:        5.321 ms (0.16% GC)
  maximum time:     9.114 ms (0.00% GC)
  --------------
  samples:          940
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

Using ForwardDiff v0.2.2:

```julia
julia> c = ForwardDiff.Chunk{10}();

julia> @time ForwardDiff.hessian!(out, rosenbrock, x, c);
 17.425611 seconds (2.38 M allocations: 60.944 MB, 0.21% gc time)
```

So it seems compilation times have improved significantly since #137 was filed.