@KristofferC 2016-12-03
Should these checks have been removed?

```jl
julia> ForwardDiff.GradientConfig{15}(rand(15))
------ AssertionError ------------------ Stacktrace (most recent call last)

AssertionError: N <= CHUNK_THRESHOLD
```